Add the ability to specify user for the spec file
Currently the code defaults to **user: daemon** and you have to manually edit the
spec to change this.  This patch introduces a `--user` flag for the
`spec` command allowing a user to be specified. No validation is
performed that this is a valid user in the container's `/etc/passwd`
file, but an error is generated by libcontainer if the user is not
found.

Docker-DCO-1.1-Signed-off-by: Phil Estes <estesp@linux.vnet.ibm.com> (github: estesp)