added a spec.Mount UnmarshalJSON to allow fstab like strings
As a long-term system administrator working with plain configuration files, I like to read and write configurations as compact as possible. If a compact format is well-known and it covers 90% of modern use cases, I recommend to support it.
This patch introduces one-line mount objects that look like the well known fstab format.

Signed-off-by: Ingo Krabbe <ikrabbe.ask@gmail.com>