README: Document pre-meeting agenda alteration
On Wed, Aug 05, 2015 at 11:57:59AM -0700, Mrunal Patel wrote [1]:
> We could also have guests on the hangout if they have some important
> topic to present. We can decide that on the topics being discussed
> in the mailing list.

On Sat, Aug 08, 2015 at 08:23:21AM -0400, Vincent Batts wrote [2]:
> We said the topics would be proposed on this list. Discussion notes have
> been shared as a Google doc so far. I'm not opposed to markdown notes, but
> initially that just seems like an repo to me.

More generally, the topic of feedback loops came up in the 2015-08-05
meeting [3,4] (after 26:02 in the video), and the consensus was to
start discussion anything that seemed worth discussion on the mailing
list ("conversations and discussions should be on
dev@opencontainers.org mailing-list first, conversations and
discussions should be on dev@opencontainers.org" [3]).  That doesn't
speak to agenda-formation specifically, but it makes an official
policy of discussing most things on the mailing list, and the two
posts I quote above extend that general approach to agenda formation.

While touching this paragraph, I also re-wrapped it to match
README.md#markdown-style.

[1]: https://groups.google.com/a/opencontainers.org/d/msg/dev/JsXgi4kxkBg/Gw7GvPodDgAJ
     From: Mrunal Patel
     Subject: Re: Hangout link for today
     Date: Wed, 5 Aug 2015 11:57:59 -0700
     Message-ID: <CANEZBD6Zs5Ht8dgkvSHRHQGaVLms_kSGqCV00AJD6eFLm9hR4w@mail.gmail.com>
     Cc: dev <dev@opencontainers.org>, …
[2]: https://groups.google.com/a/opencontainers.org/d/msg/dev/41jZ9Qe7R5c/ZInajC_0DgAJ
     From: Vincent Batts
     Subject: Re: Open Container Weekly Meeting - Aug 5, 2015
     Date: Sat, 8 Aug 2015 08:23:21 -0400
     Message-ID: <CAN6Zp5yFX8yLG3b-82SAq7AmCxVpoy1tyt0K1ijFqSsCjKPRpg@mail.gmail.com>
     Cc: dev <dev@opencontainers.org>, …
[3]: https://docs.google.com/document/d/1a5UW7MRLVaUDEjuQmRudYMZcPV0bBtPD2QEOLT_3zi0/edit?usp=sharing
[4]: https://plus.google.com/events/cqfpicicbnra9mv6kvpj0mb24u4

Signed-off-by: W. Trevor King <wking@tremily.us>