config: Add example additionalGids
The field is optional, but it's nice to have at least one example of
it in use.  The GIDs I've chosen are currently "tty" and "disk" on
Gentoo (1 is "bin"), which may be remotely reasonable choices, but the
values we're using don't really matter without an example filesystem
to provide context.

Signed-off-by: W. Trevor King <wking@tremily.us>