Cgroups: check for cpuset.cpus before set the value
Assume we put some CPUs offline and then bring them online again, since
there is not a mechanism that to help docker to refresh the state of CPUs,
all the following containers will never use those off-and-on CPUs again.

So we should always check whether there has already a change of CPUs before
we set it to the container.

Signed-off-by: Hu Keping <hukeping@huawei.com>