Deduplicate the field of RootfsPropagation
There are two RootfsPropagation fields, one is Linux.RootfsPropagation,
the other one is LinuxRuntime.RootfsPropagation. They are duplicated,
one of them should be removed.

The RootfsPropagation is definitely a runtime specific configuration,
so we remove the one of Linux.RootfsPropagation.

And the description of it is moved from config-linux.md to
runtime-config-linux.md.

Signed-off-by: Lai Jiangshan <jiangshanlai@gmail.com>