implementations: Link to kunalkushwaha/octool
Kunal mentioned this tool in this thread [1].  There's some discussion
with David Liang of OCT in that thread where Kunal and David seem to
be trying to find common ground and avoid duplicating effort, but both
projects seem to be moving along on their own now.  I think we should
link to both, and we can always remove a link later if/when the two
projects get together to provide a common validator framework.

Entries in this list (the first in implementations.md with two entries
in a category) are alphabetical by whatever the first text is in the
enter.

[1]: https://groups.google.com/a/opencontainers.org/forum/#!topic/dev/B2qQ9itIsgI
     Message-Id: <b8554ca7-bc0a-480c-9603-8c4ab3f4fdd6@opencontainers.org>

Signed-off-by: W. Trevor King <wking@tremily.us>