Warn instead of error on Seccomp config if disabled
Before this commit, passing a config with a Seccomp configuration (generated by
default by 'runc spec') to a runc binary without Seccomp support compiled in
resulted in an error on container init noting that the given Seccomp config was
being ignored. This demotes the error to a warning, to enable non-Seccomp
configurations to work out of the box.

Signed-off-by: Matthew Heon <mheon@redhat.com>