Add Architecture field to Seccomp configuration in Linux runtime
By default, Seccomp filters will only permit syscalls to be made using the
native architecture of the kernel. This is fine for most use cases, but breaks
others (such as running 32-bit code in a container on a host with a 64-bit
kernel). This patch adds a field to specify additional architectures which may
make syscalls.

Signed-off-by: Matthew Heon <mheon@redhat.com>