glossary: Specify UTF-8 for all our JSON
I wish there was a cleaner reference for what UTF-8 was.  But [1]
seems too glib, and I can't find a more targetted link than just
dropping folks into a Unicode chapter (which is what [1] does):

  The Unicode Standard, Version 6.0, §3.9 D92, §3.10 D95 (2011)

With the current v8.0 (2015-06-17), it's still §3.9 D92 and §3.10 D95.

I'd rather put this normative requirement in the configuration-spec
files, but maintainer consensus was to put it in the glossary [2,3].

[1]: https://en.wikipedia.org/wiki/UTF-8
[2]: https://github.com/opencontainers/specs/pull/146#issuecomment-138970417
[3]: https://github.com/opencontainers/specs/pull/146#issuecomment-143348788

Signed-off-by: W. Trevor King <wking@tremily.us>