config: linux: update description of PidsLimit
Fix a misleading comment for how PidsLimit works when given a limit of
0. In the PIDs controller, a limit of 0 is identical to a limit of 1,
since it is not possible to impose a limit on 0 processes.

As such, it makes no sense to distinguish the two values, rather the
value 0 (which is also the default value of an integer in Go) should
instead indicate no limit (which is the default for all new PIDs
controllers).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>