Wrap Linux 4.3 pids cgroup
With Linux 4.3 the new 'pids' cgroup becomes available. It allows to set
fork limits on a group of processes. Once a process's PID has been added to
the cgroup, all its children will automatically be added to this cgroup.

The new cgroup has the following files:

cgroups.proc: To set the processes on which to impose a fork limit;
              The entry also shows the processes that are under the fork limit

pids.max: The maximum number of processes that can be spawned;
          The file may either container the string 'max' or an unsigned
          integer value

pids.current: The number of processes in this cgroup

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>