Change the behavior when cgroupsPath is absent
The former definition was join "/" when `cgroupsPath` is absent, it's
not clear whether to join the root cgroup or create a sub cgroup under "/".

Join root cgroup would be a bad idea cause that's not what a container
should do. And craete a sub cgroup would be missing definition about what
the cgroup name should that be.

So I think we should leave this to implementations what the default
cgroup path should be.

Signed-off-by: Qiang Huang <h.huangqiang@huawei.com>