lifecycle: Don't require /run/opencontainer/<runtime>/containers
We already require it for Linux/Unix-based systems [1], so we don't
have to repeat it here.  And other systems will use different paths,
which we haven't specified yet.  When I asked why we didn't specify a
path for Windows [2], Vincent said we were waiting on help from PoC
implementations [3].  So this commit punts the location to the "State"
section, and lets the "Lifecycle" section just focus on when the
write-to-filesystem happens.

There's also discussion about removing the filesystem state registry
completely [4], in which case we'd want to remove the whole line from
the lifecycle.

[1]: https://github.com/opencontainers/specs/commit/7713efc1be66dda08548ad12da6509fd93662308#diff-b84a8d65d8ed53f4794cd2db7e8ea731L7
[2]: https://github.com/opencontainers/specs/pull/211#discussion_r41066673
[3]: https://github.com/opencontainers/specs/pull/211#discussion_r41067134
[4]: https://groups.google.com/a/opencontainers.org/forum/#!topic/dev/q6TYqVZOcX8
     Subject: removal of /run/opencontainer/containers
     Date: Wed, 25 Nov 2015 14:29:35 +0000
     Message-ID: <CAD2oYtNipt3i_C6=J4Bc-jwauo5YAvKXUqTROnPNP3vZ9+C5Vw@mail.gmail.com>

Signed-off-by: W. Trevor King <wking@tremily.us>