README: Add a link to the IRC logs
We have these irclog2html.py logs because we're part of a MeetBot pool
managed by the Linux Foundation [1], so they should be pretty
reliable.

We could have this link in the /topic, but there's already a lot there
and folks may want to browse the logs over HTTP without bothering to
connect to the live channel.

[1]: http://ircbot.wl.linuxfoundation.org/eavesdrop/%23opencontainers/%23opencontainers.2015-12-29.log.html#t2015-12-29T21:51:30

Signed-off-by: W. Trevor King <wking@tremily.us>